+++ PoporUI/Classes/UITextField/UITextField+pFormat.m
@@ -19,7 +19,7 @@
     [self formatChinaPhoneGapWidth:6];
 }
 
-- (void)formatChinaPhoneGapWidth:(int)gapWidth {
+- (void)formatChinaPhoneGapWidth:(NSInteger)gapWidth {
     UITextRange *selectedTextRange = self.selectedTextRange;
     
     NSMutableAttributedString * att = [NSMutableAttributedString separateText:self.text bigGap:gapWidth smallGap:0 separateNumberArray:@[@2, @6, @10]];
@@ -33,7 +33,7 @@
     [self formatChinaIdcardGapWidth:6];
 }
 
-- (void)formatChinaIdcardGapWidth:(int)gapWidth {
+- (void)formatChinaIdcardGapWidth:(NSInteger)gapWidth {
     UITextRange *selectedTextRange = self.selectedTextRange;
     
     NSMutableAttributedString * att = [NSMutableAttributedString separateText:self.text bigGap:gapWidth smallGap:0 separateNumberArray:@[@5, @9, @13, @17]];
@@ -43,11 +43,11 @@
 }
 
 // money
-- (void)formatMoneyUnit:(int)unit {
+- (void)formatMoneyUnit:(NSInteger)unit {
     [self formatMoneyUnit:unit gapWitdh:6];
 }
 
-- (void)formatMoneyUnit:(int)unit gapWitdh:(int)gapWidth {
+- (void)formatMoneyUnit:(NSInteger)unit gapWitdh:(NSInteger)gapWidth {
     UITextRange *selectedTextRange = self.selectedTextRange;
     
     NSMutableAttributedString * att = [NSMutableAttributedString separateMoneyText:self.text bigGap:gapWidth smallGap:0 separateNumber:unit];
@@ -56,11 +56,11 @@
     self.selectedTextRange = selectedTextRange;
 }
 
-- (void)formatBankUnit:(int)unit {
+- (void)formatBankUnit:(NSInteger)unit {
     [self formatBankUnit:unit gapWitdh:6];
 }
 
-- (void)formatBankUnit:(int)unit gapWitdh:(int)gapWidth {
+- (void)formatBankUnit:(NSInteger)unit gapWitdh:(NSInteger)gapWidth {
     UITextRange *selectedTextRange = self.selectedTextRange;
     
     NSMutableAttributedString * att = [NSMutableAttributedString separateText:self.text bigGap:gapWidth smallGap:0 separateNumber:unit];
